Implement ChatApi sending and automatic routing (#24)
Based on the number where the message was received, we
send responses via Twilio or ChatApi, automatically.

We fixed the validation of Twilio requests, and also
fixed the body parsing which was incorrectly assuming
that Twilio would be sending JSON, which is not the case.

The body sent by Twilio is also UrlEncoded, so we encode
the ChatApi body too.

Also, for whatever reason, ChatApi is re-entrant when
we send messages too, so we had to account for that
specifically too.

Finally, it turned out that none of our `Contracts`
where really needed since only public classes must
validate for null. Making everything internal makes
the warnings about validating for non-null parameters
go away.