Switch to partition per entity design
Given there are no limits to the number of partitions in a table account,
and that the partition key is what enables entity group transactions for
consistency, it's best if we move to one partition per entity (donor or
donee) so that eventually other data persisted as part of them can also
be transactionally consistent.

The rowkey now instead denotes the type of data a given row contains
within the same entity partition.