Add generic serializer based on Json.NET
The built-in `System.Text.Json` does not offer the ability to deserialize using constructors, which is a showstopper since it would force us to provide parameterless constructor for deserialization and then defeating the nullability checks we could otherwise be performing.

Also, it made it impossible to implement immutable entities, which are more important for the EventGrid work that's coming next.