Add first end-to-end feature scenario
The `Inbox.feature` file (requires the SpecFlow extension installed in order to get syntax highlighting in VS) describes the first scenario we want to support: registering a donee.

The Gherkin format allows for easy reading and understanding what the scenario is about.

The steps implementations were base intentionally generic so that they can be reused for other scenarios too, but not too generic just yet.

The scenario and steps basically run on top and verify events flowing through the `IEventStream`. This stream in production is implemented by the `EventGridStream` which will eventually push events to EventGrid, but during testing the stream just invokes any corresponding `IEventHandler<TEvent>`s that are registered for the pushed events. Combined with functions implementing that interface too, we shortcircuit the Push to EventGrid > EventGrid invokes Function > Function handles event.

Event handlers can also produce more events by pushing in turn to the stream. In particular, the `Inbox.cs` function handles `MessageReceived` (the one sent by the initial step in the scenario), does minimal augmentation and pushes semantic events `TextMessageReceived` or `ImageMessageReceived`, as well as optionally setting the national identifier if it can be determined.

Subsequently, the `Text.cs` function/handler processes it and just satisfies the behavior in the first scenario, by responding with the expected message.

We also tweaked the `IPersonRepository` so that it now provides two distinct functions: `FindAsync` can return null since it tries to locate a person by its phone number (existing behavior we had). But now once a message has been augmented with the `NationalId` for a person, retrieving it back from the repository should not involve another phone>id lookup. Also, it would not be right to return null under any circumstances from that operation (the new `GetAsync`) since at this point there must be a registered person.

Finally, `Startup` was extended to support being invoked from tests so that the DI configuration can be ported over to `Autofac` for use in tests without having to re-write them and maintain them in two places later on. Whenver we need DI from tests, we can simply new-up the `TestContainer` and resolve instances from it.

The DI configuration is also validated once at the beginning of a test run, so that we can ensure that all our types can actually be resolved and their dependencies are properly registered.