Move feature tests to their own project
This will help convey that feature tests are not unit
tests. Also removed the skipping of feature test runs
since we want to also run these on CI. We added the
secrets that are now passed as envvars when running
tests so that the minimal variables needed are in
place.