Renamed incoming/outgoing functions
Our functions will be called mostly via webhooks (i.e. chat-api/twilio/slack) or eventgrid. In the former, it's obvious that a function will receive incoming messages, but that's not so obvious in the latter: the function callbacks might be for the webhooks which send incoming messages through the grid, or they might be for messages that are to be sent back out (i.e. via messaging or slack).

So naming the classes according to whether they are incoming or outgoing processors makes it easier to tell them apart.