Added support for persisting domain entity and its events
The surface API is mostly the same, with the addition that domain
entities can be loaded in readonly mode for faster retrieval in
snapshot-style. We use the built-in support in StreamStone for
persisting extra entities with events to perform the "linear
projection".

Read-only objects cannot generate new events and will throw
an exception, but an IsReadOnly property was added so this can
be checked before interacting that way.

The `StartupWorkflow` was updated to use the new approach, and
we added a sample Donate/Donated event to showcase how state
mutation and event tracking could work in this scenario.

We're still missing the event-grid pushing of events, but it
seems that that functionality will need to also leverage the
context (as an `IPersonRepository` decorator) of the current
domain object so that when pushing to the grid, the identifier
for the owning domain object can be provided as the event
grid subject perhaps.