Simplify repository dependencies by exporting CloudStorageAccount
Instead of parsing and depending on IEnvironment for table storage-
based components, we instead just export the singleton `CloudStorageAccount`
directly, which we switch automatically to the development one for
local runs.

NOTE: local Azure Storage Emulator needs to be installed and running if the
local storage is used at all (i.e. when running the function in development
mode).