Added general-purpose POCO persistence to Tables
Having to create mirror entities inheriting from `TableEntity` was
annoying. So this new repository leverages the build-in support for
`DynamicTableEntity` which we use to persist public properties of
a POCO with a convention-based PartitionKey/RowKey pair: the former
is the full type name of the entity, the latter, the property
annotated with `[Key]` from `System.ComponentModel.DataAnnotations`.

When rehydrating the POCO, we use the same "trick" we use for the
domain object project: we write the properties as a bit of json,
then let the deserializer create it. This will properly invoke the
right constructor, so that POCOs don't really need to be poorly
encapsulated DTOs.