Add event grid decorator for person repository
The decorator pushes to the event grid (via the event stream)
whenever a Person is persisted. The event stream was simplified
and also extended so that (optional) event metadata can be provided
when pushing events, which we use from the repository decorator
to set the event subject to the person id.

In order to more fully exercise the end to end testing, we also
switched to using the actual instances of all storage-based interfaces,
which now requires the local azure storage emulator in order to run
them. But this makes for a more realistic acceptance testing via
the specflow docs.