Simplify event pushing and add storing of all eventgrid events
In order to simplify the event stream API, I went back to the original
signature without an explicit EventMetadata object, and instead just
provide it as a (mostly explicitly implemented) interface that the
events themselves can use to hint the conversion from the payload to
an event grid event.

Also introduced a separate topic "path" for `Domain` events and other
things we might push to eventgrid (topic `System` for now). Seems like
we'd want to persist the former and not the latter, so that can be
tweaked in the new `save-event` azure function subscription.

Also added feature scenario skipping via a simple tag, and added a
Donate.feature to showcase that (also to see events persisted locally
while it run and failed locally once ;)).

Added the robust JoinableTaskContext/Factory as an alternative to
doing `.Wait()` where we need to, since that has known problems and
can cause deadlocks.

Finally, a new `[NoExport]` attribute is introduced to flag types
that we don't want in the DI container for whatever reason (i.e.
the domain event types implement an interface now, which would
make them candidates for being in DI, but that makes no sense).