Move all behavior code to a new library project
Due to a bug in the Azure Functions SDK, Go To Definition and even
most intellisense is completely broken, making the entire experience
terrible. So we just implement the workaround mentioned in the bug
at https://github.com/Azure/azure-functions-vs-build-sdk/issues/344
to place all code in a library instead.

The Api project now only contains the azure functions and nothing
more. It calls into the DI-injected dependencies as necessary, in
particular, it now consumes the IEventHandler<T> instead of implementing
those itself.