Drastically simply config/secrets by leveraging ASP.NET Core configuration
Turns out that there is a built-in way to just add stuff to a configuration
aggregate that can even support Azure KeyVault natively. So we removed all
our previous stuff and `kv-` prefixes, and added the plain variables there.

For the local settings, we check it in but don't copy it to output in CI.
Approach pretty much copied from https://www.tomfaltesek.com/azure-functions-local-settings-json-and-source-control/.

Devs need to create a `secrets.settings.json` (which is in `.gitignore` to
avoid accidentally checking it in) with the following format and keys:

```
{
  "AZURE_CLIENT_ID": "...",
  "AZURE_CLIENT_SECRET": "...",
  "AZURE_TENANT_ID": "..."
}
```

And with that, unit and acceptance tests can be run locally,
as well as F5 on the Api project.