Add tax id recognition and verification
We validate that that person has a valid tax status.
Added CommandBus for easier invocation of this tax
validation behavior, so as not to force a new event
to be raised (which isn't really what's going on).

The startup workflow now tries to validate the tax
status upon registration.

Pending sending welcome and notice messages if validation
results in the person not being eligible for donations.

Pending batch processing when AFIP site is down.