Simplify and unify the event grid handlers
By generalizing the invocation of the IEventHandler<T> for
various message types, we make the EventGrid "facade"
much simpler and easier to maintain (and reason about).

It's now clearer than ever that we're just using the EventGrid
as a persistent/dispatcher/retry mechanism, while all the
logic lives in Core.