Simplify messaging API, system number is implied
Rather than persisting and passing around the To/From pair
constantly to know through which messaging channel to
reply to a user, we just keep another map (PhoneSystem)
that is updated constantly whenever users contact the
system. This becomes their "From" number for replies we
need to send them, automatically, without having to keep
the original number they sent messages to.

This was becoming important as we decoupled domain
events which only had a reference to the PersonId, from
which we could just get the PhoneNumber. So rather than
adding the SystemPhone/To everywhere, we instead just
simplify everything.