Centralize, unify and simplify slack messaging usage
Rather than having various event handlers decide how to
send messages to slack, we make the same design choice
as with messaging: we create a `SlackMessageSent`  and
have it sent via the EventGrid, where a single handler deals
with it and how it's sent, in a uniform way (i.e. detemrining
channel and thread id to use).