Add automation pause/resume behavior
Whenever a user starts interacting with support via slack,
the automated processing via the workflows/pipeline is
wasteful and could even result in unexpected replies from
the system when the user is expecting a response from (or
responding to) a human. So we add a unified mechanism
to all messages sent to slack to offer the Pause/Resume
actions, which quickly set the given phone number as
non-automated, meaning all messages incoming for it are
forwarded as-is to Slack (to the given ongoing thread) and
no further processing is performed.

Note that this is decoupled from a registered Person since
we might want to switch to this mode when helping a new
user get registered in the first place.