Add explicit registration via Slack UI
When a registration fails, we report all attempted images to Slack and
let the user know we'll follow up. The new Register action in Slack
allows explicitly filling-in the user information from what we can read
manually from the received images. When submitting this information,
we leverage QR code generation to re-trigger the flow entirely, as if
a new image was received that can be properly recognized this time.

In order to simplify the increasing number of interactions with Slack,
I moved all the code into Core\Slack, which now implement their
own interface for processors, which the two HTTP hooks now call
in a generic fashion. Each processor determines whether a message
applies to it before doing processing. This massively reduces the
complexity of the former switch/ifs we had to determine what sort
of processing needed to happen.

Also, all these processors do is trigger more events thorugh the
EventGrid.