Fix renamed property so deserialization works again
The EventGrid.Data > type deserialization leverages Json.NET ability
to invoke the right constructor passing the value from the payload.
In this case, the payload value was `PhoneNumber` (since that's the
name of the property in the base class) but the constructor parameter
remained `to`, therefore it couldn't be matched and resulted in a null
value being passed.