Add payment processing events and functions
We're still not sending the PaymentApproved since
that component is still missing, but this already
implements the end-to-end with MP.

We extended the event grid handler dispatching to
allow event handlers to throw an `HttpStatusException`
to signal a specific HTTP status code to return to EventGrid.

The `PaymentCodeReceived` uses this to trigger auto-retry
for QR codes that still don't have a payment associated.