Add tests and docs to durable action
Since DurableAction has its own table, we use the action name as the
partition key (which makes it easier to read and should partition
better)  instead of the action ID (which can be a guid or some other instance identifier).

Improve the entity cleanup to only do it on sucesss/cancel.
We were previously always deleting the entity up-front, and then
potentially updating it on a retry. This means a retried operation would
would always incur three trips to storage: initial retrieval, initial
deletion, and finally its update. Now that would be two instead.

In addition, this adds some additional resiliency for the case where the
function is shutdown before the method execution finishes: previously
this would have caused the previous retry attempts to be lost since it
would have been deleted on action start. We'd now preserve it.