Add concurrency detection for person repository
We previously had "last save wins", which wasn't a very good idea.
By preserving the original version of the domain object we read,
we can pass that to the event store when writing, so an exception
is thrown if the underlying version changed before we saved.

We'll then ensure this never happens in production by leveraging
durable entity functions so that invocations to each person entity
happens in a serial fashion.