Simplify batch status validation and add unit tests
The concept of commands was very confusing and didn't add any
value, since it was used in a single place that had very little
code that could be moved up to the timer-triggered function too
with minimal effort, saving along the way two classes and an
entire new concept to grasp in the app architecture.

Therefore, `TaxStatusValidation` is now the entity we store for
signaling the pending batch validations, and the corresponding
`TaxStatusValidationHandler` performs the actual handling of
the `PersonRegistered` event to do immediate validation of a
single person, which, upon succeding, will delete the validation
entry.

If it doesn't succeed, the subsequent timer-triggered function
will re-attempt that at a later time again.