Make the EventType fixed so we can subscribe from logic apps
The Logic Apps UI doesn't allow (that I could find easily)
subscribing to event types by wildcard or "starts-with"
conditions, which would make it almost impossible to get all
exceptions.

On top of that, the Topic field (which could have been used
to filter all `Runtime` events) is not even surfaced in the
subscription UI either, so we can't use that (see
https://feedback.azure.com/forums/287593-logic-apps/suggestions/39786736-add-the-topic-field-to-the-event-grid-domain-trigg).

So the easiest was to just make the event type constant. The
full exception information will be part of the Data of the
event anyway, so we can still see what specific exception it
was.