Rename PhoneSystem > PhoneEntry
This better expresses that instances of this class are entries
in our "phone directory" of users who are interacting with the
system via a phone number (i.e. WhatsApp).

This phone entry also allows us to persist additional information
about the user, such as his intended role in the system, even
if we still haven't fully registered them. The new `Role` prop
will allow us to simplify the `StartupWorkflow` which won't
need to accomodate two simulateous registration workflows (for
Donor and Donee), since we will use this `Role` property in the
phone entry to switch workflows too (in addition to the registeed
person Role only). Given this new mechanism, we'll also need to
pass in the `PhoneEntry` to workflows, so we make it part of
the `IWorkflow` API too.